KOGITO-5072: Isolate fetch calls.
