+++ packages/dmn-dev-sandbox-form-webapp/src/DmnDevSandboxAppDataApi.tsx
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { DmnFormSchema } from "@kogito-tooling/form/dist/dmn";
+
+export interface AppData {
+  filename: string;
+  modelName: string;
+  schema: DmnFormSchema;
+  formUrl: string;
+  modelUrl: string;
+  swaggerUIUrl: string;
+}
+
+export async function fetchAppData(): Promise<AppData> {
+  const response = await fetch("/data.json");
+  return (await response.json()) as AppData;
+}
